+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/views/ViewManagerDefinition.kt
@@ -4,16 +4,11 @@ import android.content.Context
 import android.view.View
 import android.view.ViewGroup
 import com.facebook.react.bridge.ReactContext
-import com.facebook.react.bridge.ReadableMap
 import expo.modules.adapters.react.NativeModulesProxy
 import expo.modules.core.ViewManager
 import expo.modules.kotlin.AppContext
-import expo.modules.kotlin.DynamicNull
 import expo.modules.kotlin.exception.CodedException
-import expo.modules.kotlin.exception.toCodedException
 import expo.modules.kotlin.functions.BaseAsyncFunctionComponent
-import expo.modules.kotlin.logger
-import expo.modules.kotlin.recycle
 
 class ViewManagerDefinition(
   private val viewFactory: (Context, AppContext) -> View,
@@ -35,29 +30,6 @@ class ViewManagerDefinition(
       ViewManager.ViewManagerType.GROUP
     } else {
       ViewManager.ViewManagerType.SIMPLE
-    }
-  }
-
-  fun setProps(propsToSet: ReadableMap, onView: View) {
-    props.forEach { (name, propDelegate) ->
-      try {
-        if (propsToSet.hasKey(name)) {
-          propsToSet
-            .getDynamic(name)
-            .recycle {
-              propDelegate.set(this, onView)
-            }
-        } else if (propDelegate.isNullable) {
-          propDelegate.set(DynamicNull, onView)
-        }
-      } catch (exception: Throwable) {
-        logger.error("❌ Cannot set the '$name' prop on the '${viewType.simpleName}'", exception)
-
-        handleException(
-          onView,
-          exception.toCodedException()
-        )
-      }
     }
   }